Revert "Fix Code improvement"
This reverts commit 7edda09fcfffc74359c012203073e413632757ed.
