--- .gitignore
@@ -1,7 +1,0 @@
-venv_scrapy
-__pycache__
-.env
-.vscode
-.pyc
-db.sqlite
-log_error.log
