+++ crawl_jobs/database_setup.py
@@ -1,7 +1,7 @@
-from decouple import config
 from sqlalchemy import create_engine
-
 from crawl_jobs.models import Base
+from decouple import config
 
-engine = create_engine(config('DATABASE'), echo=True)
+
+engine = create_engine(config("DATABASE"), echo=False)
 Base.metadata.create_all(engine)
