+++ crawl_jobs/models.py
@@ -1,10 +1,8 @@
+from sqlalchemy import Column, String, Integer, DateTime
+from sqlalchemy.ext.declarative import declarative_base
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Integer, String
-from sqlalchemy.ext.declarative import declarative_base
-
 Base = declarative_base()
-
 
 class Job(Base):
     __tablename__ = 'jobs'
