--- readme.md
@@ -1,28 +1,0 @@
-# Search for jobs on the Pyjobs website
-
-## Before you start:
-
-You need the version ```Python 3.7```
-
-1) See available commands:
-   ```shell
-    $ make help
-   ```
-
-1) Installation requirements:
-   ```shell
-    $ make requirements
-   ```
-
-## Run application:
-   ```shell
-    $ make run
-   ```
-
-   The data will be stored in sqlite.
-
-- For example: 
-
-| id | name | Company | Salary | State | Local | Level | Description |
-| :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
-| 1 | Arquiteto de Dados Nível Sênior – Barueri/SP | Empresa: Hrsoul | Faixa salarial: R$ 10.000,01 a R$ 13.000,00 | Estado: São Paulo | Local da vaga: Barueri | Nível: Sênior | Imprescindível sólida experiência com arquitetura, Oracle, MSSQL, Informix, Vertica, Hive, MongoDB, MySQL, PostgreSQL, Hadoop, DB Cloud, Talend, Elastic, Solr e REDIS; | Desejáveis conhecimentos em NoSQL e Docker; | Graduação em TI ou áreas relacionadas; | Disponibilidade para atuação em Alphaville – São Paulo. | Requisitos | Imprescindível sólida experiência com arquitetura, Oracle, MSSQL, Informix, Vertica, Hive, MongoDB, MySQL, PostgreSQL, Hadoop, DB Cloud, Talend, Elastic, Solr e REDIS; | Desejáveis conhecimentos em NoSQL e Docker; | Graduação em TI ou áreas relacionadas; | Disponibilidade para atuação em Alphaville – São Paulo.
