--- requirements.txt
@@ -1,3 +1,0 @@
-python-decouple==3.3
-Scrapy==2.0.1
-SQLAlchemy==1.3.16
