--- requirements_dev.txt
@@ -1,5 +1,0 @@
--r requirements.txt
-autopep8==1.5.2
-black==19.10b0
-ipdb==0.13.2
-isort==4.3.21
